API-1489: move common/components/Table/ExpandableTableRow to common/components/ExpandableTableRow
