+++ src/Akeneo/Connectivity/Connection/front/src/common/components/ExpandableTableRow.tsx
