+++ src/Akeneo/Connectivity/Connection/front/tests/src/common/components/ExpandableTableRow.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import '@testing-library/jest-dom/extend-expect';
 import {screen} from '@testing-library/react';
-import ExpandableTableRow from '@src/common/components/Table/ExpandableTableRow';
-import {renderWithProviders} from '../../../../test-utils';
+import ExpandableTableRow from '@src/common/components/ExpandableTableRow';
+import {renderWithProviders} from '../../../test-utils';
 import userEvent from '@testing-library/user-event';
 
 describe('testing expandable table row', () => {
